Implement writer controller and refactor routes for article management
